+++ NEWS.md
@@ -5,3 +5,7 @@
 * Second release of FishPhyloMaker - errors in insertions fixed (v0.1.2)
 
 * A new function `whichFishAdd` was released
+
+* A stable version of FishPhyloMaker is now on CRAN
+
+* Function PD_deficit now outputs total and inserted phylogenetic information